+++ src/app/constants/styles.ts
@@ -3,7 +3,7 @@ export const BUTTON_STYLE = {
   loginW: (className: string) =>
     `w-[700px] h-20 text-[26px] text-[#486283] bg-white font-semibold rounded-[30px] ${className}`,
   loginB: (className: string) =>
-    `w-[700px] h-20 text-[26px] text-white bg-[#486283] font-semibold rounded-[30px] ${className} `,
+    `w-[700px] ${className} h-20 text-[26px] text-white bg-[#486283] font-semibold rounded-[30px]`,
   auth: (className: string) =>
     `w-[220px] h-[70px] text-[26px] text-[#486283] bg-[#CED7E4] font-semibold rounded-[30px] ${className} `,
   complete: (className: string) =>
