+++ src/app/utils/toast.ts
@@ -1,15 +1,15 @@
-import { toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
+import { toast } from 'react-toastify'
+import 'react-toastify/dist/ReactToastify.css'
 
 export const editAlert = () => {
-  toast.info('수정되었습니다', {
-    icon: () => '✍️',
+  toast.info('5분내에 인증코드를 입력해주세요', {
+    icon: () => '🔑',
     theme: 'dark',
     closeOnClick: true,
-    autoClose: 2000,
+    autoClose: 1500,
     position: 'top-center',
-  });
-};
+  })
+}
 
 export const deleteAlert = () => {
   toast.error('삭제되었습니다', {
@@ -18,8 +18,8 @@ export const deleteAlert = () => {
     closeOnClick: true,
     autoClose: 2000,
     position: 'top-center',
-  });
-};
+  })
+}
 
 export const exceedAlert = () => {
   toast.warning('최대 2명까지 추가 가능합니다', {
@@ -28,8 +28,8 @@ export const exceedAlert = () => {
     closeOnClick: true,
     autoClose: 2000,
     position: 'top-center',
-  });
-};
+  })
+}
 
 export const overlapAlert = () => {
   toast.warning('회원은 중복해서 추가할 수 없습니다.', {
@@ -38,8 +38,8 @@ export const overlapAlert = () => {
     closeOnClick: true,
     autoClose: 2000,
     position: 'top-center',
-  });
-};
+  })
+}
 
 export const previewAlert = () => {
   toast.info('더블클릭으로 작성으로 돌아가기', {
@@ -48,8 +48,8 @@ export const previewAlert = () => {
     closeOnClick: true,
     autoClose: 2000,
     position: 'top-center',
-  });
-};
+  })
+}
 
 export const thumbnailAlert = () => {
   toast.warning('썸네일을 입력해주세요', {
@@ -58,8 +58,8 @@ export const thumbnailAlert = () => {
     closeOnClick: true,
     autoClose: 2000,
     position: 'top-center',
-  });
-};
+  })
+}
 
 export const noContentsAlert = () => {
   toast.info('본문에 최소 1개의 컨텐츠를 추가해주세요', {
@@ -68,8 +68,8 @@ export const noContentsAlert = () => {
     closeOnClick: true,
     autoClose: 2000,
     position: 'top-center',
-  });
-};
+  })
+}
 
 export const deletePostAlert = () => {
   toast.error('게시물이 삭제되었습니다. 잠시후 메인페이지로 이동합니다.', {
@@ -78,8 +78,8 @@ export const deletePostAlert = () => {
     closeOnClick: true,
     autoClose: 2000,
     position: 'top-center',
-  });
-};
+  })
+}
 
 export const textLimitAlert = () => {
   toast.warning('텍스트는 255자 이하로 작성해주세요', {
@@ -88,5 +88,5 @@ export const textLimitAlert = () => {
     closeOnClick: true,
     autoClose: 2000,
     position: 'top-center',
-  });
-};
+  })
+}
